Change SendReport & sendDate "id" values from 16 to 8bits
The underlying HID_::sendReport method is anyhow only transmitting the first byte of the "id" field, so the content of the second byte is currently ignored. One can then just as well change the "id" argument type from 16 to 8bits.

Update the HIDPowerDevice_::sendDate the same way, since it's just passing on the "id" argument to HID_::sendReport.

+++ src/HID/HID.h
@@ -115,7 +115,7 @@ class HID_ : public PluggableUSBModule
 public:
     HID_(void);
     int begin(void);
-    int SendReport(uint16_t id, const void* data, int len);
+    int SendReport(uint8_t id, const void* data, int len);
     int SetFeature(uint16_t id, const void* data, int len);
     bool LockFeature(uint16_t id, bool lock);
